Refatorados pacotes que antes eram globais e na verdade são da feature episodes.
