+++ app/src/main/java/com/example/renerd/app/MainModule.kt
@@ -1,6 +1,6 @@
 package com.example.renerd.app
 
-import com.example.renerd.components.floating_player.di.FloatingPlayerModule
+import com.example.renerd.features.episodes.components.floating_player.di.FloatingPlayerModule
 import com.example.renerd.features.episodes.di.EpisodesModule
 import com.example.renerd.features.player.di.PlayerModule
