+++ app/src/main/java/com/example/renerd/features/episodes/components/filters_dialog/tabs/FiltersProductTabFragment.kt
@@ -1,4 +1,4 @@
-package com.example.renerd.components.filters_dialog.tabs
+package com.example.renerd.features.episodes.components.filters_dialog.tabs
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,15 +6,15 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.renerd.components.filters_dialog.FilterTabListener
-import com.example.renerd.components.filters_dialog.adapters.FilterItemAdapter
+import com.example.renerd.features.episodes.components.filters_dialog.FilterTabListener
+import com.example.renerd.features.episodes.components.filters_dialog.adapters.FilterItemAdapter
 import com.example.renerd.databinding.CLayoutFilterTabBinding
 import com.example.renerd.view_models.FiltersTabsItemModel
 
 
-class FiltersGuestTabFragment(
-    private val guestsList: List<FiltersTabsItemModel>,
-    private val filterTabListener: FilterTabListener
+class FiltersProductTabFragment(
+    private val productsList: List<FiltersTabsItemModel>,
+    private val filterTabListener: FilterTabListener,
 ) : Fragment() {
 
     private lateinit var binding: CLayoutFilterTabBinding
@@ -24,7 +24,6 @@ class FiltersGuestTabFragment(
         savedInstanceState: Bundle?
     ): View {
         binding = CLayoutFilterTabBinding.inflate(inflater, container, false)
-
 
         return binding.root
     }
@@ -44,7 +43,7 @@ class FiltersGuestTabFragment(
 
 
     private fun setUpTitle(){
-        binding.title.text = "Filtro de Convidados"
+        binding.title.text = "Filtro de Produtores"
     }
 
 
@@ -53,7 +52,7 @@ class FiltersGuestTabFragment(
         binding.recyclerviewBase.layoutManager = LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false)
 
         val filterItemAdapter = FilterItemAdapter(
-            filtersTabsListItemModel = guestsList,
+            filtersTabsListItemModel = productsList,
             onClick = { filtersTabsListItemModel ->
                 filterTabListener.onItemValeuChange(filtersTabsListItemModel)
             }
@@ -63,10 +62,9 @@ class FiltersGuestTabFragment(
     }
 
 
-
     companion object {
-        fun newInstance(guestsList: List<FiltersTabsItemModel>, filterTabListener: FilterTabListener):FiltersGuestTabFragment {
-            return FiltersGuestTabFragment(guestsList, filterTabListener)
+        fun newInstance(productsList: List<FiltersTabsItemModel>, filterTabListener: FilterTabListener): FiltersProductTabFragment {
+            return FiltersProductTabFragment(productsList, filterTabListener)
         }
     }
 }
