+++ app/src/main/java/com/example/renerd/features/episodes/components/filters_dialog/tabs/FiltersYearTabFragment.kt
@@ -1,4 +1,4 @@
-package com.example.renerd.components.filters_dialog.tabs
+package com.example.renerd.features.episodes.components.filters_dialog.tabs
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,15 +6,15 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.renerd.components.filters_dialog.FilterTabListener
-import com.example.renerd.components.filters_dialog.adapters.FilterItemAdapter
+import com.example.renerd.features.episodes.components.filters_dialog.FilterTabListener
+import com.example.renerd.features.episodes.components.filters_dialog.adapters.FilterItemAdapter
 import com.example.renerd.databinding.CLayoutFilterTabBinding
 import com.example.renerd.view_models.FiltersTabsItemModel
 
 
-class FiltersSubjectTabFragment(
-    private val subjectsList: List<FiltersTabsItemModel>,
-    private val filterTabListener: FilterTabListener,
+class FiltersYearTabFragment(
+    private val yearsList: List<FiltersTabsItemModel>,
+    private val filterTabListener: FilterTabListener
 ) : Fragment() {
 
     private lateinit var binding: CLayoutFilterTabBinding
@@ -24,6 +24,7 @@ class FiltersSubjectTabFragment(
         savedInstanceState: Bundle?
     ): View {
         binding = CLayoutFilterTabBinding.inflate(inflater, container, false)
+
 
         return binding.root
     }
@@ -40,16 +41,21 @@ class FiltersSubjectTabFragment(
         this.setUpRecyclerView()
     }
 
+
+
+
     private fun setUpTitle(){
-        binding.title.text = "Filtro de Temas"
+        binding.title.text = "Filtro por Ano"
     }
+
+
 
 
     private fun setUpRecyclerView(){
         binding.recyclerviewBase.layoutManager = LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false)
 
         val filterItemAdapter = FilterItemAdapter(
-            filtersTabsListItemModel = subjectsList,
+            filtersTabsListItemModel = yearsList,
             onClick = { filtersTabsListItemModel ->
                 filterTabListener.onItemValeuChange(filtersTabsListItemModel)
             }
@@ -58,9 +64,12 @@ class FiltersSubjectTabFragment(
         binding.recyclerviewBase.adapter = filterItemAdapter
     }
 
+
+
+
     companion object {
-        fun newInstance(subjectsList: List<FiltersTabsItemModel>, filterTabListener: FilterTabListener):FiltersSubjectTabFragment {
-            return FiltersSubjectTabFragment(subjectsList, filterTabListener)
+        fun newInstance(yearsList: List<FiltersTabsItemModel>, filterTabListener: FilterTabListener): FiltersYearTabFragment {
+            return FiltersYearTabFragment(yearsList, filterTabListener)
         }
     }
 }
