+++ app/src/main/java/com/example/renerd/features/episodes/components/floating_player/FloatingPlayerPresenter.kt
@@ -1,4 +1,4 @@
-package com.example.renerd.components.floating_player
+package com.example.renerd.features.episodes.components.floating_player
 
 
 import com.example.renerd.view_models.EpisodeViewModel
@@ -7,7 +7,8 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 
-class FloatingPlayerPresenter(private val repository: FloatingPlayerContract.Repository): FloatingPlayerContract.Presenter {
+class FloatingPlayerPresenter(private val repository: FloatingPlayerContract.Repository):
+    FloatingPlayerContract.Presenter {
     private var view: FloatingPlayerContract.View? = null
 
     override fun attachView(view: FloatingPlayerContract.View) {
